Extracted util functions to seperate js file. Fixed bugs with click events on icons
